Rename the item kind file.
